modify `active` structure
